First version beta docs
it's done first version to integration backend
