+++ src/Components/Forms/index.js
@@ -1,4 +1,4 @@
-export { default as BasicForm } from "./BasicForm";
+export { default as Forms } from "./Form";
 export { default as Register } from "./Register";
 export { default as FormOnlySelect } from "./FormOnlySelect";
 export { default as FormDisable } from "./FormDisable";
