+++ src/Page/Authentication/Authentication.jsx
@@ -1,3 +1,4 @@
+import PropTypes from "prop-types";
 import { useState } from "react";
 import { Container } from "react-bootstrap";
 import { SubtitleWithBadge, Paragraph } from "@Components/Title";
@@ -7,7 +8,7 @@ import { RenderJson } from "@Components/Json";
 
 import { ReturnJson } from "./Utils";
 
-const Authentication = () => {
+const Authentication = ({ id }) => {
   const [show, setShow] = useState("Body");
 
   const handleChange = (e) => {
@@ -16,40 +17,46 @@ const Authentication = () => {
 
   return (
     <>
-      <SubtitleWithBadge title="Authentication" className="text-dark" badge="POST" />
-      <Paragraph>For can to call every resources you need to be authenticated.</Paragraph>
-      <FormDisable
-        data={[
-          {
-            controlId: "disableTextId",
-            label: "EndPoint",
-            control: "text",
-            placeholder: window.location + "auth",
-          },
-        ]}
-        className="mb-2"
-      />
-      <FormOnlySelect
-        data={[
-          {
-            value: "Body",
-            text: "Body",
-          },
-          {
-            value: "Preview",
-            text: "Preview",
-          },
-        ]}
-        state={show}
-        handleChange={handleChange}
-      />
-      <Code className={"mt-2"}>
-        <Container>
-          <RenderJson src={ReturnJson(show)} settings={{ name: false }} />
-        </Container>
-      </Code>
+      <section id={id}>
+        <SubtitleWithBadge title="Authentication" className="text-dark" badge="POST" />
+        <Paragraph>For can to call every resources you need to be authenticated.</Paragraph>
+        <FormDisable
+          data={[
+            {
+              controlId: "disableTextId",
+              label: "EndPoint",
+              control: "text",
+              placeholder: window.location + "auth",
+            },
+          ]}
+          className="mb-2"
+        />
+        <FormOnlySelect
+          data={[
+            {
+              value: "Body",
+              text: "Body",
+            },
+            {
+              value: "Preview",
+              text: "Preview",
+            },
+          ]}
+          state={show}
+          handleChange={handleChange}
+        />
+        <Code className={"mt-2"}>
+          <Container>
+            <RenderJson src={ReturnJson(show)} settings={{ name: false }} />
+          </Container>
+        </Code>
+      </section>
     </>
   );
+};
+
+Authentication.propTypes = {
+  id: PropTypes.string,
 };
 
 export default Authentication;
