+++ src/Page/Header/Utils/ReturnDataForm.utils.js
@@ -1,0 +1,28 @@
+const Form = [
+  {
+    type: "select",
+    label: "",
+    controlId: "formBasicRequest",
+    selectSettings: {
+      ariaLabel: "Choose request type",
+      options: [
+        {
+          value: "Embarcaciones",
+          text: "Embarcaciones",
+        },
+        {
+          value: "Users",
+          text: "Users",
+        },
+        {
+          value: "Sales",
+          text: "Sales",
+        },
+      ],
+    },
+  },
+];
+
+const ReturnDataForm = () => Form;
+
+export default ReturnDataForm;
