+++ src/Page/index.js
@@ -4,3 +4,4 @@ export { default as GetUnit } from "./GetUnit/GetUnit";
 export { default as Post } from "./Post/Post";
 export { default as Put } from "./Put/Put";
 export { default as Delete } from "./Delete/Delete";
+export { default as Header } from "./Header/Header";
